requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
Appium-Python-Client==2.11.1
attrs==23.1.0
certifi==2022.6.15.1
charset-normalizer==3.3.2
click==8.1.7
coverage==7.6.1
driver==0.1.0
exceptiongroup==1.1.3
h11==0.14.0
idna==3.4
iniconfig==2.0.0
outcome==1.2.0
packaging==23.1
peewee==3.17.6
pip==22.0.4
pluggy==1.3.0
psycopg2-binary==2.9.9
PySocks==1.7.1
pytest-cov==2.12.1
pytest==7.4.2
rand==0.3.0
requests==2.32.3
selenium==4.12.0
setuptools==58.1.0
sniffio==1.3.0
sortedcontainers==2.4.0
toml==0.10.2
tomli==2.0.1
trio-websocket==0.10.4
trio==0.22.2
urllib3==2.0.4
wsproto==1.2.0
